# API-Gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing LinguaLeap System resources",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/progress/achievements": {
            "get": {
                "description": "Get the user's achievements",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get achievements",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.AchievementsResponse"
                        }
                    }
                }
            }
        },
        "/progress/daily": {
            "get": {
                "description": "Get the user's daily progress",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get daily progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.DailyProgressResponse"
                        }
                    }
                }
            }
        },
        "/progress/delete-goal/{goal_id}": {
            "delete": {
                "description": "Delete an existing goal",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Delete a goal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Goal ID",
                        "name": "goal_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.GoalResponse"
                        }
                    }
                }
            }
        },
        "/progress/goals": {
            "get": {
                "description": "Get the user's goals",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get goals",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.GoalsResponse"
                        }
                    }
                }
            }
        },
        "/progress/leaderboard/{language_code}": {
            "get": {
                "description": "Get the leaderboard for a specific language",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get leaderboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language Code",
                        "name": "language_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.LeaderboardResponse"
                        }
                    }
                }
            }
        },
        "/progress/monthly": {
            "get": {
                "description": "Get the user's monthly progress",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get monthly progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.MonthlyProgressResponse"
                        }
                    }
                }
            }
        },
        "/progress/set-goal": {
            "post": {
                "description": "Set a new goal for the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Set a goal",
                "parameters": [
                    {
                        "description": "Set Goal",
                        "name": "goal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/progres.Goal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.GoalResponse"
                        }
                    }
                }
            }
        },
        "/progress/skills/{language_code}": {
            "get": {
                "description": "Get the user's skills for a specific language",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get skills",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language Code",
                        "name": "language_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.SkillsResponse"
                        }
                    }
                }
            }
        },
        "/progress/statistics/{language_code}": {
            "get": {
                "description": "Get the user's statistics for a specific language",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language Code",
                        "name": "language_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.StatisticsResponse"
                        }
                    }
                }
            }
        },
        "/progress/update-goal/{goal_id}": {
            "put": {
                "description": "Update an existing goal for the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Update a goal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Goal ID",
                        "name": "goal_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Goal",
                        "name": "goal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/progres.UpdateGoalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.GoalResponse"
                        }
                    }
                }
            }
        },
        "/progress/weekly": {
            "get": {
                "description": "Get the user's weekly progress",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get weekly progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.WeeklyProgressResponse"
                        }
                    }
                }
            }
        },
        "/progress/{language_code}": {
            "get": {
                "description": "Get the user's progress for a specific language",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get language progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language Code",
                        "name": "language_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/progres.ProgressResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "progres.Achievement": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Yutuq yaratilgan sana",
                    "type": "string"
                },
                "description": {
                    "description": "Yutuq ta'rifi",
                    "type": "string"
                },
                "earned_at": {
                    "description": "Yutuq erishilgan sana",
                    "type": "string"
                },
                "id": {
                    "description": "Yutuq ID'si",
                    "type": "string"
                },
                "title": {
                    "description": "Yutuq nomi",
                    "type": "string"
                },
                "user_id": {
                    "description": "Foydalanuvchi ID'si",
                    "type": "string"
                }
            }
        },
        "progres.AchievementsResponse": {
            "type": "object",
            "properties": {
                "achievements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/progres.Achievement"
                    }
                }
            }
        },
        "progres.DailyProgress": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Progress yaratilgan sana",
                    "type": "string"
                },
                "date": {
                    "description": "Sana",
                    "type": "string"
                },
                "id": {
                    "description": "Kunlik progress ID'si",
                    "type": "string"
                },
                "lessons_completed": {
                    "description": "Bajarilgan darslar soni",
                    "type": "integer"
                },
                "minutes_practiced": {
                    "description": "Mashq qilingan daqiqalar",
                    "type": "integer"
                },
                "new_words_learned": {
                    "description": "O'rganilgan yangi so'zlar soni",
                    "type": "integer"
                },
                "streak_days": {
                    "description": "Kunlik davomiylik kunlari",
                    "type": "integer"
                },
                "user_id": {
                    "description": "Foydalanuvchi ID'si",
                    "type": "string"
                },
                "xp_earned": {
                    "description": "Kunlik XP miqdori",
                    "type": "integer"
                }
            }
        },
        "progres.DailyProgressResponse": {
            "type": "object",
            "properties": {
                "progress": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/progres.DailyProgress"
                    }
                }
            }
        },
        "progres.Goal": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Maqsad yaratilgan sana",
                    "type": "string"
                },
                "current_value": {
                    "description": "Hozirgi qiymat",
                    "type": "integer"
                },
                "end_date": {
                    "description": "Maqsad tugash sanasi",
                    "type": "string"
                },
                "goal_type": {
                    "description": "Maqsad turi (masalan, 'daily_xp', 'weekly_lessons', va hokazo)",
                    "type": "string"
                },
                "id": {
                    "description": "Maqsad ID'si",
                    "type": "string"
                },
                "language_code": {
                    "description": "Til kodi (masalan, 'fr', 'es', va hokazo)",
                    "type": "string"
                },
                "start_date": {
                    "description": "Maqsad boshlanish sanasi",
                    "type": "string"
                },
                "target_value": {
                    "description": "Maqsad uchun kerakli qiymat",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "Maqsad yangilangan sanaa",
                    "type": "string"
                },
                "user_id": {
                    "description": "Foydalanuvchi ID'si",
                    "type": "string"
                }
            }
        },
        "progres.GoalResponse": {
            "type": "object",
            "properties": {
                "goal": {
                    "$ref": "#/definitions/progres.Goal"
                }
            }
        },
        "progres.GoalsResponse": {
            "type": "object",
            "properties": {
                "goals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/progres.Goal"
                    }
                }
            }
        },
        "progres.Leaderboard": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Jadval yaratilgan sana",
                    "type": "string"
                },
                "id": {
                    "description": "Liderlar jadvali ID'si",
                    "type": "string"
                },
                "language_code": {
                    "description": "Til kodi",
                    "type": "string"
                },
                "level": {
                    "description": "Daraja",
                    "type": "string"
                },
                "rank": {
                    "description": "O'rin",
                    "type": "integer"
                },
                "user_id": {
                    "description": "Foydalanuvchi ID'si",
                    "type": "string"
                },
                "xp": {
                    "description": "XP miqdori",
                    "type": "integer"
                }
            }
        },
        "progres.LeaderboardResponse": {
            "type": "object",
            "properties": {
                "leaderboards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/progres.Leaderboard"
                    }
                }
            }
        },
        "progres.MonthlyProgress": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Progress yaratilgan sana",
                    "type": "string"
                },
                "id": {
                    "description": "Oylik progress ID'si",
                    "type": "string"
                },
                "lessons_completed": {
                    "description": "Bajarilgan darslar soni",
                    "type": "integer"
                },
                "month": {
                    "description": "Oy",
                    "type": "string"
                },
                "most_improved_skill": {
                    "description": "Eng ko'p rivojlangan ko'nikma",
                    "type": "string"
                },
                "new_words_learned": {
                    "description": "O'rganilgan yangi so'zlar soni",
                    "type": "integer"
                },
                "total_minutes_practiced": {
                    "description": "Mashq qilingan umumiy daqiqalar",
                    "type": "integer"
                },
                "total_xp_earned": {
                    "description": "Oylik umumiy XP miqdori",
                    "type": "integer"
                },
                "user_id": {
                    "description": "Foydalanuvchi ID'si",
                    "type": "string"
                }
            }
        },
        "progres.MonthlyProgressResponse": {
            "type": "object",
            "properties": {
                "progress": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/progres.MonthlyProgress"
                    }
                }
            }
        },
        "progres.ProgressResponse": {
            "type": "object",
            "properties": {
                "daily_progress": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/progres.DailyProgress"
                    }
                },
                "monthly_progress": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/progres.MonthlyProgress"
                    }
                },
                "weekly_progress": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/progres.WeeklyProgress"
                    }
                }
            }
        },
        "progres.Skill": {
            "type": "object",
            "properties": {
                "comprehension_rate": {
                    "description": "Tinglab tushunish darajasi",
                    "type": "integer"
                },
                "created_at": {
                    "description": "Ko'nikma yaratilgan sana",
                    "type": "string"
                },
                "fluency_rate": {
                    "description": "So'zlashuv ravonligi darajasi",
                    "type": "integer"
                },
                "id": {
                    "description": "Ko'nikma ID'si",
                    "type": "string"
                },
                "language_code": {
                    "description": "Til kodi",
                    "type": "string"
                },
                "level": {
                    "description": "Ko'nikma darajasi",
                    "type": "integer"
                },
                "name": {
                    "description": "Ko'nikma nomi (masalan, 'Vocabulary', 'Grammar', 'Listening', va hokazo)",
                    "type": "string"
                },
                "progress_to_next_level": {
                    "description": "Keyingi darajaga o'tish uchun progress",
                    "type": "integer"
                },
                "rules_mastered": {
                    "description": "Egallangan qoidalar soni",
                    "type": "integer"
                },
                "user_id": {
                    "description": "Foydalanuvchi ID'si",
                    "type": "string"
                },
                "words_learned": {
                    "description": "O'rganilgan so'zlar soni",
                    "type": "integer"
                }
            }
        },
        "progres.SkillsResponse": {
            "type": "object",
            "properties": {
                "skills": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/progres.Skill"
                    }
                }
            }
        },
        "progres.Statistic": {
            "type": "object",
            "properties": {
                "accuracy_rate": {
                    "description": "Aniqlik darajasi (%)",
                    "type": "integer"
                },
                "average_daily_time": {
                    "description": "O'rtacha kunlik o'qish vaqti (daqiqalarda)",
                    "type": "integer"
                },
                "created_at": {
                    "description": "Statistika yaratilgan sana",
                    "type": "string"
                },
                "id": {
                    "description": "Statistika ID'si",
                    "type": "string"
                },
                "language_code": {
                    "description": "Til kodi",
                    "type": "string"
                },
                "most_challenging_skill": {
                    "description": "Eng qiyin ko'nikma",
                    "type": "string"
                },
                "strongest_skill": {
                    "description": "Eng kuchli ko'nikma",
                    "type": "string"
                },
                "total_exercises_completed": {
                    "description": "Bajarilgan umumiy mashqlar soni",
                    "type": "integer"
                },
                "total_learning_time": {
                    "description": "Umumiy o'qish vaqti (daqiqalarda)",
                    "type": "integer"
                },
                "total_lessons_completed": {
                    "description": "Bajarilgan umumiy darslar soni",
                    "type": "integer"
                },
                "user_id": {
                    "description": "Foydalanuvchi ID'si",
                    "type": "string"
                }
            }
        },
        "progres.StatisticsResponse": {
            "type": "object",
            "properties": {
                "statistics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/progres.Statistic"
                    }
                }
            }
        },
        "progres.UpdateGoalRequest": {
            "type": "object",
            "properties": {
                "goal": {
                    "$ref": "#/definitions/progres.Goal"
                },
                "goal_id": {
                    "type": "string"
                }
            }
        },
...
        "progres.WeeklyProgressResponse": {
            "type": "object",
            "properties": {
                "progress": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/progres.WeeklyProgress"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}